c4ads-rag-iran-twitter/pyproject.toml
[tool.poetry]
name = "ai-demo"
version = "0.1.0"
description = ""
authors = ["jdelorimier <james.delorimier@c4ads.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.10.16"
python-dotenv = "^1.0.1"
llama-index-readers-file = "^0.1.8"
llama-index-vector-stores-milvus = "^0.1.5"
chromadb = "^0.4.24"
llama-index-vector-stores-chroma = "^0.1.5"
llama-index-storage-kvstore-redis = "^0.1.2"
streamlit = "^1.32.0"
ruff = "^0.3.2"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
